requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.35.0
langchain==0.2.5
langchain-openai==0.1.9
langchain-community==0.2.5
faiss-cpu==1.13.1
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.4
slowapi==0.1.9
